requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-decouple==3.8
openai==1.3.8
pytesseract==0.3.10
pdfplumber
Pillow==10.1.0
pydantic[email]==2.5.0
PyMuPDF
bcrypt==4.0.1
PyPDF2==3.0.1
PyYAML==6.0.2
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
tika==1.24.0  
requests==2.32.4